Move ResourceDeleter to external_resources_manager.cpp
