+++ src/execution/operator/helper/physical_external_resource.cpp
@@ -1,7 +1,6 @@
 #include "duckdb/execution/operator/helper/physical_external_resource.hpp"
 
 #include "duckdb/execution/operator/helper/launch_external_resource.hpp"
-#include "duckdb/main/attached_database.hpp"
 #include "duckdb/main/client_context.hpp"
 #include "duckdb/main/database.hpp"
 #include "duckdb/main/external_resources_manager.hpp"
