+++ src/main/attached_database.cpp
@@ -1,12 +1,10 @@
 #include "duckdb/main/attached_database.hpp"
 #include "duckdb/logging/log_manager.hpp"
-#include "duckdb/logging/log_type.hpp"
 
 #include "duckdb/catalog/duck_catalog.hpp"
 #include "duckdb/common/constants.hpp"
 #include "duckdb/common/enums/checkpoint_on_detach.hpp"
 #include "duckdb/common/file_system.hpp"
-#include "duckdb/main/connection.hpp"
 #include "duckdb/main/database.hpp"
 #include "duckdb/main/database_manager.hpp"
 #include "duckdb/main/settings.hpp"
@@ -344,49 +342,6 @@ void AttachedDatabase::OnDetach(ClientContext &context) {
 	}
 	if (stored_database_path && visibility != AttachVisibility::HIDDEN) {
 		stored_database_path->OnDetach();
-	}
-}
-
-ResourceDeleter::ResourceDeleter(DatabaseInstance &db, string deleter_function_p, Value deleter_payload_p,
-                                 string resource_type_p, string resource_name_p)
-    : db(db), deleter_function(std::move(deleter_function_p)), deleter_payload(std::move(deleter_payload_p)),
-      resource_type(std::move(resource_type_p)), resource_name(std::move(resource_name_p)) {
-}
-
-string ResourceDeleter::DeleteSQL() const {
-	if (deleter_function.empty()) {
-		return string();
-	}
-	// Render the (possibly schema-qualified) function name with each component properly quoted, so a
-	// name needing quoting is valid SQL and a registry-supplied string cannot inject SQL.
-	auto function_name = QualifiedName::Parse(deleter_function).ToString();
-	return "SELECT * FROM " + function_name + "(" + deleter_payload.ToSQLString() + ")";
-}
-
-void ResourceDeleter::Delete() {
-	auto sql = DeleteSQL();
-	if (sql.empty()) {
-		return;
-	}
-	// On a separate internal connection, since the current connection's context lock is held.
-	Connection con(db);
-	auto result = con.Query(sql);
-	DUCKDB_LOG(db, ExternalResourceLogType, resource_type, resource_name, string("destroy"),
-	           result->HasError() ? result->GetError() : string(),
-	           Value::MAP(LogicalType::VARCHAR, LogicalType::VARCHAR, vector<Value>(), vector<Value>()));
-	if (result->HasError()) {
-		// A failed teardown is a leak, so fail loudly and say how to retry it manually.
-		throw IOException("external resource teardown failed: %s. The resource was NOT torn down; run `%s;` "
-		                  "to retry the teardown",
-		                  result->GetError(), sql);
-	}
-}
-
-void ResourceDeleter::TryDelete() {
-	try {
-		Delete();
-	} catch (std::exception &ex) {
-		DUCKDB_LOG_WARNING(db, ErrorData(ex).Message());
 	}
 }
